Rename utils.py
