+++ test/components/preprocessors/test_nltk_document_splitter.py
@@ -5,7 +5,7 @@ from haystack import Document
 from pytest import LogCaptureFixture
 
 from haystack.components.preprocessors.nltk_document_splitter import NLTKDocumentSplitter
-from haystack.components.preprocessors.utils import SentenceSplitter
+from haystack.components.preprocessors._nltk_document_splitter_utils import SentenceSplitter
 
 
 def test_init_warning_message(caplog: LogCaptureFixture) -> None:
